frames_getter/requirements.txt
# RabbitMQ
amqp==5.1.0

# Process
regex==2023.5.5
python-dotenv==1.0.0
pyyaml==6.0
requests==2.31.0
numpy==1.24.3
opencv-python-headless==4.7.0.72